# Conflicts:
#	.idea/dataSources.xml
#	src/main/resources/com/example/project7/editer_quiz/_1_SelectionTypeDeProjet.fxml
#	src/main/resources/com/example/project7/editer_quiz/_2_EditerProjet.fxml
#	src/main/resources/com/example/project7/editer_quiz/_4_EditerQCM.fxml
#	src/main/resources/com/example/project7/editer_quiz/_5_EditerQCU.fxml
#	src/main/resources/com/example/project7/editer_quiz/_6_EditerQuestionLibre.fxml
#	src/main/resources/com/example/project7/editer_quiz/_7_EditerDescription.fxml
#	src/main/resources/com/example/project7/editer_quiz/_8_SupprimerSection.fxml

+++ src/main/java/com/example/project7/model/QuestionLibre.java
@@ -2,7 +2,7 @@ package com.example.project7.model;
 
 public class QuestionLibre  extends Section{
     private String question;
-    private int scoreTotal;
+    private float scoreTotal;
     private int nombreScore;
     private int nombreLigne;
     private float tailleLigne;
@@ -25,11 +25,11 @@ public class QuestionLibre  extends Section{
         this.question = question;
     }
 
-    public int getScoreTotal() {
+    public float getScoreTotal() {
         return scoreTotal;
     }
 
-    public void setScoreTotal(int scoreTotal) {
+    public void setScoreTotal(float scoreTotal) {
         this.scoreTotal = scoreTotal;
     }
